+++ server/core/initializers/migrations/1085-video-channel-sync-full-sync-cutoff.ts
@@ -1,0 +1,24 @@
+import * as Sequelize from 'sequelize'
+
+async function up (utils: {
+  transaction: Sequelize.Transaction
+  queryInterface: Sequelize.QueryInterface
+  sequelize: Sequelize.Sequelize
+}): Promise<void> {
+  const { transaction } = utils
+
+  await utils.queryInterface.addColumn('videoChannelSync', 'fullSyncCutoffAt', {
+    type: Sequelize.DATE,
+    allowNull: true,
+    defaultValue: null
+  }, { transaction })
+}
+
+function down () {
+  throw new Error('Not implemented.')
+}
+
+export {
+  down,
+  up
+}